requirements.txt
mwapi
mwoauth
pywikibot
pyyaml
requests_oauthlib
